microservice/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pandas
python-dotenv
mysql-connector-python